new stages

+++ recruit-me-ui/src/web/src/app/modules/candidat/table-list/table-list.component.ts
@@ -6,6 +6,9 @@ import { Candidat } from "../../../core/models/candidat";
 import { Router } from '@angular/router';
 import {environment} from "../../../../environments/environment";
 import {AuthenticationService} from "../../../login/auth.service";
+import {MatDialog} from "@angular/material/dialog";
+import {YesOrNOComponent} from "../../../shared/yes-or-no/yes-or-no.component";
+import {MatSnackBar} from "@angular/material/snack-bar";
 
 
 @Component({
@@ -19,7 +22,7 @@ export class TableListComponent implements OnInit {
     downloadUrl = environment.baseUrlCandidat+'candidats/file/';
 
     constructor(private candidatService: CandidatService, private authService: AuthenticationService,
-                private router: Router) { }
+                private router: Router, public dialog: MatDialog, private _snackBar: MatSnackBar) { }
 
     ngOnInit() {
         console.log('ngOnInit');
@@ -56,13 +59,27 @@ export class TableListComponent implements OnInit {
     }
 
     deleteCandidat(id: number) {
-        this.candidatService.deleteCandidat(id)
-            .subscribe(
-                data => {
-                    console.log(data);
-                    this.reloadData();
-                },
-                error => console.log(error));
+        const nomUtilisateur = this.candidats.filter(candidat => candidat.id === id)[0].nomUtilisateur;
+        const dialogRef = this.dialog.open(YesOrNOComponent, {
+            width: '250px',
+            data: {message: `you want to delete ${nomUtilisateur} ?`}
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+                this.candidatService.deleteCandidat(id)
+                    .subscribe(
+                        data => {
+                            console.log(data);
+                            this._snackBar.open("utilisateur supprimé", "OK", {
+                                duration: 2000,
+                            });
+                            this.reloadData();
+                        },
+                        error => console.log(error));
+            }
+        });
+
     }
 
     //candidatDetails(id: number){
